Move Test for aborted metadata extraction to `TestExtractDsMeta`
